float to double

+++ app/src/main/java/com/trembleturn/trembleturn/POJO/EndLocation.java
@@ -7,8 +7,8 @@ import com.google.gson.annotations.SerializedName;
 public class EndLocation{
 	
 	@SerializedName("lat")
-	public float lat;
+	public double lat;
 
 	@SerializedName("lng")
-	public float lng;
+	public double lng;
 }
